Show single purr
It's funny. If I had build up my tests a little bit differently I would
have first build to show a single Purr and have a component for that,
and then loop over it for the PurrStream. But I went for the array
straight away throwing me of the TDD game. :P